Allow consumeFirstStringMessageFrom to be called multiple times (#61)
In v0.5.0 you were able to call consumeFirstStringMessageFrom multiple
times, with each invocation causing the next message in a topic to be
consumed. In v0.10.0 this was no longer the case: the first invocation
works fine, but subsequent invocations would result in exceptions as the
first invocation would cause *all* messages to be committed even though
only the first one is returned, leaving no messages remaining to be
consumed by subsequent calls. This regression was likely introduced by
commit a90674307eb15206e4f75619beae067124d989d6.

This commit fixes this by disabling auto-commit by default and simply
committing only the message that consumeFirstStringMessageFrom returns.